% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{computeIntersection}
\alias{computeIntersection}
\title{Find the intersections of two functions}
\usage{
computeIntersection(a, b, c, d)
}
\arguments{
\item{a}{A numeric.}

\item{b}{A numeric.}

\item{c}{A numeric.}

\item{d}{A numeric.}
}
\value{
The intersection of the two functions or NA if there is not one.
}
\description{
Find the intersections of two functions
}
