% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{computeLinearFunction}
\alias{computeLinearFunction}
\title{Compute the function value f(x) of a function}
\usage{
computeLinearFunction(x, a, b)
}
\arguments{
\item{x}{A numeric.}

\item{a}{A numeric.}

\item{b}{A numeric.}
}
\value{
The function value f(x) of a function.
}
\description{
Compute the function value f(x) of a function
}
