% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{drawMinOrMaxLine}
\alias{drawMinOrMaxLine}
\title{Draw the line of the max or min of all functions}
\usage{
drawMinOrMaxLine(matrix, minmax)
}
\arguments{
\item{matrix}{A matrix.}

\item{minmax}{A string.}
}
\value{
The data frame of the segments of the min or max function.
}
\description{
Draw the line of the max or min of all functions
}
