% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{generateColors}
\alias{generateColors}
\title{Generates random colors for the linear functions}
\usage{
generateColors(numOfRows)
}
\arguments{
\item{numOfRows}{A integer.}
}
\value{
The colors for the linear functions.
}
\description{
Generates random colors for the linear functions
}
