% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{generateFunctionName}
\alias{generateFunctionName}
\title{Generates name for the linear function}
\usage{
generateFunctionName(i, matrix)
}
\arguments{
\item{i}{A integer.}

\item{matrix}{A matrix.}
}
\value{
The name of the linear function as a string.
}
\description{
Generates name for the linear function
}
