% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{generateMatrixFromSaddleVector}
\alias{generateMatrixFromSaddleVector}
\title{Generates a mx2 matrix from the saddlepoints vector.}
\usage{
generateMatrixFromSaddleVector(vector)
}
\arguments{
\item{vector}{A vector.}
}
\value{
The matrix of saddlepoints formated mx2 from \code{vector}.
}
\description{
Generates a mx2 matrix from the saddlepoints vector.
}
