% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getAllIntersections}
\alias{getAllIntersections}
\title{Find all intersections of the functions}
\usage{
getAllIntersections(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\value{
The function values of the functions intersections in sorted vector.
}
\description{
Find all intersections of the functions
}
