% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getMaxOfRowMin}
\alias{getMaxOfRowMin}
\title{Find the maximum of the rows minima.}
\usage{
getMaxOfRowMin(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
The maximum of the rows minima as numeric.
}
\description{
Find the maximum of the rows minima.
}
