% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getMinOfColMax}
\alias{getMinOfColMax}
\title{Find the minimum of the columns maxima.}
\usage{
getMinOfColMax(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
The minimum of the columns maxima as numeric.
}
\description{
Find the minimum of the columns maxima.
}
