% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getMinOrMaxFunctionsValueForIntersections}
\alias{getMinOrMaxFunctionsValueForIntersections}
\title{Find the min or max functions value for each intersection}
\usage{
getMinOrMaxFunctionsValueForIntersections(matrix, intersections, minmax)
}
\arguments{
\item{matrix}{A matrix.}

\item{intersections}{A vector.}

\item{minmax}{A string.}
}
\value{
The vector of the intersection min/max function value tupel.
}
\description{
Find the min or max functions value for each intersection
}
