% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getSaddlePointsOfGame}
\alias{getSaddlePointsOfGame}
\title{Find the saddlepoints of the game.}
\usage{
getSaddlePointsOfGame(matrix, maxCol)
}
\arguments{
\item{matrix}{A matrix}

\item{maxCol}{A numeric}
}
\value{
The matrix of saddlepoints.
}
\description{
Find the saddlepoints of the game.
}
