% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{isMatrixOfRightDimention}
\alias{isMatrixOfRightDimention}
\title{Check if the matrix has the right format of 2xn or mx2}
\usage{
isMatrixOfRightDimention(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\value{
The TRUE if the matrix is of the right format or FALSE if not.
}
\description{
Check if the matrix has the right format of 2xn or mx2
}
