% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plotSolution}
\alias{plotSolution}
\title{Plot the graphical solution of a 2xn or mx2 matrixgame}
\usage{
plotSolution(matrix, rowOrCol = "row")
}
\arguments{
\item{matrix}{A matrix.}

\item{rowOrCol}{A string.}
}
\value{
The plot of the solution in a cartesian coordinate system with a legend.
}
\description{
Plot the graphical solution of a 2xn or mx2 matrixgame
}
