% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{solveGame}
\alias{solveGame}
\title{Solve a two-person zero-sum game.}
\usage{
solveGame(matrix)
}
\arguments{
\item{matrix}{A matrix.}
}
\value{
The value of the game \code{matrix}, the saddlepoints of the \code{matrix} if existing and the strategies of the players if the strategies are not pure.
}
\description{
Solve a two-person zero-sum game.
}
\examples{
solveGame(matrix(c(10,2,2,10,4,8,1,12), ncol = 2, byrow = TRUE))
solveGame(matrix(c(-3,1,2,0,5,2,3,2,2,-3,1,-3,1,0,-2,2), nrow = 4, byrow = TRUE))
}
