% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{solveLinearProgram}
\alias{solveLinearProgram}
\title{Solve the linear program of the gamematrix}
\usage{
solveLinearProgram(matrix, minmax)
}
\arguments{
\item{matrix}{A matrix.}

\item{minmax}{A string.}
}
\value{
The max or min value of the game and the probabilities of the players strategies.
}
\description{
Solve the linear program of the gamematrix
}
